Module Shadow_stdlib

val raise : exn -> 'a

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.raise] instead

val raise_notrace : exn -> 'a

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.raise_notrace] instead

val invalid_arg : string -> 'a

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.invalid_arg] instead

val failwith : string -> 'a

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.failwith] instead

exception Exit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Referring to the stdlib directly is discouraged by Base. You should either use the equivalent functionality offered by Base, or if you really want to refer to the stdlib, use Stdlib.Exit instead

exception Not_found

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Instead of raising [Not_found], consider using [raise_s] with an informative error message. If code needs to distinguish [Not_found] from other exceptions, please change it to handle both [Not_found] and [Not_found_s]. Then, instead of raising [Not_found], raise [Not_found_s] with an informative error message.

val (=) : 'a -> 'a -> bool

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( = )] instead

val (<>) : 'a -> 'a -> bool

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( <> )] instead

val (<) : 'a -> 'a -> bool

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( < )] instead

val (>) : 'a -> 'a -> bool

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( > )] instead

val (<=) : 'a -> 'a -> bool

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( <= )] instead

val (>=) : 'a -> 'a -> bool

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( >= )] instead

val compare : 'a -> 'a -> int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.compare] instead

val min : 'a -> 'a -> 'a

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.min] instead

val max : 'a -> 'a -> 'a

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.max] instead

val (==) : 'a -> 'a -> bool

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [phys_equal] instead.

val (!=) : 'a -> 'a -> bool

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [not (phys_equal ...)] instead.

val not : bool -> bool

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.not] instead

val (&&) : bool -> bool -> bool

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( && )] instead

val (||) : bool -> bool -> bool

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( || )] instead

val __LOC__ : string

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.__LOC__] instead

val __FILE__ : string

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.__FILE__] instead

val __LINE__ : int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.__LINE__] instead

val __MODULE__ : string

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.__MODULE__] instead

val __POS__ : string * int * int * int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.__POS__] instead

val __FUNCTION__ : string

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.__FUNCTION__] instead

val __LOC_OF__ : 'a -> string * 'a

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.__LOC_OF__] instead

val __LINE_OF__ : 'a -> int * 'a

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.__LINE_OF__] instead

val __POS_OF__ : 'a -> (string * int * int * int) * 'a

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.__POS_OF__] instead

val (|>) : 'a -> ('a -> 'b) -> 'b

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( |> )] instead

val (@@) : ('a -> 'b) -> 'a -> 'b

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( @@ )] instead

val (~-) : int -> int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( ~- )] instead

val (~+) : int -> int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( ~+ )] instead

val succ : int -> int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Int.succ] instead.

val pred : int -> int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Int.pred] instead.

val (+) : int -> int -> int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( + )] instead

val (-) : int -> int -> int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( - )] instead

val (*) : int -> int -> int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( * )] instead

val (/) : int -> int -> int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( / )] instead

val (mod) : int -> int -> int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use (%), which has slightly different semantics, or Int.rem which is equivalent.

val abs : int -> int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.abs] instead

val max_int : int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Int.max_value] instead.

val min_int : int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Int.min_value] instead.

val (land) : int -> int -> int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( land )] instead

val (lor) : int -> int -> int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( lor )] instead

val (lxor) : int -> int -> int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( lxor )] instead

val lnot : int -> int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.lnot] instead

val (lsl) : int -> int -> int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( lsl )] instead

val (lsr) : int -> int -> int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( lsr )] instead

val (asr) : int -> int -> int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( asr )] instead

val (~-.) : float -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( ~-. )] instead

val (~+.) : float -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( ~+. )] instead

val (+.) : float -> float -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( +. )] instead

val (-.) : float -> float -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( -. )] instead

val (*.) : float -> float -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( *. )] instead

val (/.) : float -> float -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( /. )] instead

val (**) : float -> float -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [**.] instead.

val sqrt : float -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.sqrt] instead.

val exp : float -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.exp] instead.

val log : float -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.log] instead.

val log10 : float -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.log10] instead.

val expm1 : float -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.expm1] instead.

val log1p : float -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.log1p] instead.

val cos : float -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.cos] instead.

val sin : float -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.sin] instead.

val tan : float -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.tan] instead.

val acos : float -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.acos] instead.

val asin : float -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.asin] instead.

val atan : float -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.atan] instead.

val atan2 : float -> float -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.atan2] instead.

val hypot : float -> float -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.hypot] instead.

val cosh : float -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.cosh] instead.

val sinh : float -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.sinh] instead.

val tanh : float -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.tanh] instead.

val acosh : float -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.acosh] instead.

val asinh : float -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.asinh] instead.

val atanh : float -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.atanh] instead.

val ceil : float -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.round_up] instead.

val floor : float -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.round_down] instead.

val abs_float : float -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.abs_float] instead

val copysign : float -> float -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.copysign] instead.

val mod_float : float -> float -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.mod_float] instead.

val frexp : float -> float * int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.frexp] instead.

val ldexp : float -> int -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.ldexp] instead.

val modf : float -> float * float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.modf] instead.

val float : int -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.of_int] instead.

val float_of_int : int -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.of_int] instead.

val truncate : float -> int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Int.of_float] instead.

val int_of_float : float -> int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Int.of_float] instead.

val infinity : float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.infinity] instead.

val neg_infinity : float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.neg_infinity] instead.

val nan : float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.nan] instead.

val max_float : float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.max_finite_value] instead.

val min_float : float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.min_positive_normal_value] instead.

val epsilon_float : float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.epsilon_float] instead.

type nonrec fpclass = fpclass =

| FP_normal

| FP_subnormal

| FP_zero

| FP_infinite

| FP_nan

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Referring to the stdlib directly is discouraged by Base. You should either use the equivalent functionality offered by Base, or if you really want to refer to the stdlib, use Stdlib.fpclass instead

val classify_float : float -> fpclass

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.classify] instead.

val (^) : string -> string -> string

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( ^ )] instead

val int_of_char : char -> int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Char.to_int] instead.

val char_of_int : int -> char

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Char.of_int_exn] instead.

val ignore : 'a -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.ignore] instead

val string_of_bool : bool -> string

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Bool.to_string] instead.

val bool_of_string_opt : string -> bool option

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.bool_of_string_opt] instead

val bool_of_string : string -> bool

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Bool.of_string] instead.

val string_of_int : int -> string

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Int.to_string] instead.

val int_of_string_opt : string -> int option

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.int_of_string_opt] instead

val int_of_string : string -> int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Int.of_string] instead.

val string_of_float : float -> string

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.to_string] instead.

val float_of_string_opt : string -> float option

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.float_of_string_opt] instead

val float_of_string : string -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Float.of_string] instead.

val fst : ('a * 'b) -> 'a

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.fst] instead

val snd : ('a * 'b) -> 'b

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.snd] instead

val (@) : 'a list -> 'a list -> 'a list

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( @ )] instead

type nonrec in_channel = in_channel

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.In_channel.t] instead.

type nonrec out_channel = out_channel

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.Out_channel.t] instead.

val stdin : in_channel

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.stdin] instead.

val stdout : out_channel

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.stdout] instead.

val stderr : out_channel

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.stderr] instead.

val print_char : char -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.Out_channel.output_char Stdio.stdout] instead.

val print_string : string -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.Out_channel.output_string Stdio.stdout] instead.

val print_bytes : bytes -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.Out_channel.output_bytes Stdio.stdout] instead.

val print_int : int -> unit

deprecated

val print_float : float -> unit

deprecated

val print_endline : string -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.print_endline] instead.

val print_newline : unit -> unit

deprecated

val prerr_char : char -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.Out_channel.output_char Stdio.stderr] instead.

val prerr_string : string -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.Out_channel.output_string Stdio.stderr] instead.

val prerr_bytes : bytes -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.Out_channel.output_bytes Stdio.stderr] instead.

val prerr_int : int -> unit

deprecated

val prerr_float : float -> unit

deprecated

val prerr_endline : string -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.prerr_endline] instead.

val prerr_newline : unit -> unit

deprecated

val read_line : unit -> string

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.In_channel.input_line] instead.

val read_int_opt : unit -> int option

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.read_int_opt] instead

val read_int : unit -> int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.read_int] instead

val read_float_opt : unit -> float option

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.read_float_opt] instead

val read_float : unit -> float

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.read_float] instead

type nonrec open_flag = open_flag =

| Open_rdonly

| Open_wronly

| Open_append

| Open_creat

| Open_trunc

| Open_excl

| Open_binary

| Open_text

| Open_nonblock

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Referring to the stdlib directly is discouraged by Base. You should either use the equivalent functionality offered by Base, or if you really want to refer to the stdlib, use Stdlib.open_flag instead

val open_out : string -> out_channel

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.Out_channel.create] instead.

val open_out_bin : string -> out_channel

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.Out_channel.create] instead.

val open_out_gen : open_flag list -> int -> string -> out_channel

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.open_out_gen] instead

val flush : out_channel -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.Out_channel.flush] instead.

val flush_all : unit -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.flush_all] instead

val output_char : out_channel -> char -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.Out_channel.output_char] instead.

val output_string : out_channel -> string -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.Out_channel.output_string] instead.

val output_bytes : out_channel -> bytes -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.Out_channel.output_bytes] instead.

val output : out_channel -> bytes -> int -> int -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.Out_channel.output] instead.

val output_substring : out_channel -> string -> int -> int -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.Out_channel.output] instead.

val output_byte : out_channel -> int -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.Out_channel.output_byte] instead.

val output_binary_int : out_channel -> int -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.Out_channel.output_binary_int] instead.

val output_value : out_channel -> 'a -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.Out_channel.output_value] instead.

val seek_out : out_channel -> int -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.Out_channel.seek] instead.

val pos_out : out_channel -> int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.Out_channel.pos] instead.

val out_channel_length : out_channel -> int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.Out_channel.length] instead.

val close_out : out_channel -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.Out_channel.close] instead.

val close_out_noerr : out_channel -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.Out_channel.close] instead.

val set_binary_mode_out : out_channel -> bool -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.Out_channel.set_binary_mode] instead.

val open_in : string -> in_channel

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.In_channel.create] instead.

val open_in_bin : string -> in_channel

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.In_channel.create] instead.

val open_in_gen : open_flag list -> int -> string -> in_channel

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.open_in_gen] instead

val input_char : in_channel -> char

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.In_channel.input_char] instead.

val input_line : in_channel -> string

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.In_channel.input_line] instead.

val input : in_channel -> bytes -> int -> int -> int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.In_channel.input] instead.

val really_input : in_channel -> bytes -> int -> int -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.In_channel.really_input] instead.

val really_input_string : in_channel -> int -> string

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is no equivalent functionality in Base or Stdio but you can use [Stdio.In_channel] instead. Alternatively, if you really want to refer the stdlib you can use [Stdlib.really_input_string].

val input_byte : in_channel -> int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.In_channel.input_byte] instead.

val input_binary_int : in_channel -> int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.In_channel.input_binary_int] instead.

val input_value : in_channel -> 'a

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.In_channel.unsafe_input_value] instead.

val seek_in : in_channel -> int -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.In_channel.seek] instead.

val pos_in : in_channel -> int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.In_channel.pos] instead.

val in_channel_length : in_channel -> int

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.In_channel.length] instead.

val close_in : in_channel -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.In_channel.close] instead.

val close_in_noerr : in_channel -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.In_channel.close] instead.

val set_binary_mode_in : in_channel -> bool -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Stdio.In_channel.set_binary_mode] instead.

module LargeFile = LargeFile

type nonrec 'a ref = 'a ref = {

mutable contents : 'a;

}

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Referring to the stdlib directly is discouraged by Base. You should either use the equivalent functionality offered by Base, or if you really want to refer to the stdlib, use Stdlib.ref instead

val ref : 'a -> 'a ref

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.ref] instead

val (!) : 'a ref -> 'a

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( ! )] instead

val (:=) : 'a ref -> 'a -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( := )] instead

val incr : int ref -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Int.incr] instead.

val decr : int ref -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Int.decr] instead.

type nonrec ('a, 'b) result = ('a, 'b) result =

| Ok of 'a

| Error of 'b

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Use [Result.t] instead.

type nonrec ('a, 'b, 'c, 'd, 'e, 'f) format6 = ('a, 'b, 'c, 'd, 'e, 'f) format6

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Referring to the stdlib directly is discouraged by Base. You should either use the equivalent functionality offered by Base, or if you really want to refer to the stdlib, use Stdlib.format6 instead

type nonrec ('a, 'b, 'c, 'd) format4 = ('a, 'b, 'c, 'd) format4

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Referring to the stdlib directly is discouraged by Base. You should either use the equivalent functionality offered by Base, or if you really want to refer to the stdlib, use Stdlib.format4 instead

type nonrec ('a, 'b, 'c) format = ('a, 'b, 'c) format

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. Referring to the stdlib directly is discouraged by Base. You should either use the equivalent functionality offered by Base, or if you really want to refer to the stdlib, use Stdlib.format instead

val string_of_format : ('a, 'b, 'c, 'd, 'e, 'f) format6 -> string

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.string_of_format] instead

val format_of_string : ('a, 'b, 'c, 'd, 'e, 'f) format6 -> ('a, 'b, 'c, 'd, 'e, 'f) format6

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.format_of_string] instead

val (^^) : ('a, 'b, 'c, 'd, 'e, 'f) format6 -> ('f, 'b, 'c, 'e, 'g, 'h) format6 -> ('a, 'b, 'c, 'd, 'g, 'h) format6

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.( ^^ )] instead

val exit : int -> 'a

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.exit] instead

val at_exit : (unit -> unit) -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.at_exit] instead

val valid_float_lexem : string -> string

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.valid_float_lexem] instead

val unsafe_really_input : in_channel -> bytes -> int -> int -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.unsafe_really_input] instead

val do_at_exit : unit -> unit

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.do_at_exit] instead

val do_domain_local_at_exit : (unit -> unit) ref

deprecated [2016-09] this element comes from the stdlib distributed with OCaml. There is not equivalent functionality in Base or Stdio at the moment, so you need to use [Stdlib.do_domain_local_at_exit] instead

module Arg = Arg

module Array = Array

module ArrayLabels = ArrayLabels

module Atomic = Atomic

module Bool = Bool

module Buffer = Buffer

module Bytes = Bytes

module BytesLabels = BytesLabels

module Callback = Callback

module Char = Char

module Complex = Complex

module Condition = Condition

module Digest = Digest

module Domain = Domain

module Dynarray = Dynarray

module Effect = Effect

module Either = Either

module Ephemeron = Ephemeron

module Filename = Filename

module Float = Float

module Format = Format

module Fun = Fun

module Gc = Gc

module Hashtbl = Hashtbl

module In_channel = In_channel

module Int = Int

module Int32 = Int32

module Int64 = Int64

module Lazy = Lazy

module Lexing = Lexing

module List = List

module ListLabels = ListLabels

module Map = Map

module Marshal = Marshal

module MoreLabels = MoreLabels

module Mutex = Mutex

module Nativeint = Nativeint

module Obj = Obj

module Oo = Oo

module Option = Option

module Out_channel = Out_channel

module Parsing = Parsing

module Printexc = Printexc

module Printf = Printf

module Queue = Queue

module Random = Random

module Result = Result

module Scanf = Scanf

module Semaphore = Semaphore

module Seq = Seq

module Set = Set

module Stack = Stack

module StdLabels = StdLabels

module String = String

module StringLabels = StringLabels

module Sys = Sys

module Type = Type

module Uchar = Uchar

module Unit = Unit

module Weak = Weak